backend/requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
python-dotenv